+++ cache/cache_test.go
@@ -28,81 +28,86 @@ var _ = Suite(&CacheSuite{})
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-func (s *CacheSuite) TestStore(c *C) {
-	store := New(time.Second/16, time.Second/32)
+func (s *CacheSuite) TestCache(c *C) {
+	cache := New(time.Second/16, time.Second/32)
 
-	store.Set("1", "TEST")
-	store.Set("2", "TEST")
+	cache.Set("1", "TEST")
+	cache.Set("2", "TEST")
 
-	c.Assert(store.Get("1"), Equals, "TEST")
-	c.Assert(store.Get("2"), Equals, "TEST")
+	c.Assert(cache.Size(), Equals, 2)
 
-	c.Assert(store.Has("2"), Equals, true)
-	c.Assert(store.Has("3"), Equals, false)
+	c.Assert(cache.Get("1"), Equals, "TEST")
+	c.Assert(cache.Get("2"), Equals, "TEST")
 
-	item, exp := store.GetWithExpiration("1")
+	c.Assert(cache.Has("2"), Equals, true)
+	c.Assert(cache.Has("3"), Equals, false)
+
+	item, exp := cache.GetWithExpiration("1")
 
 	c.Assert(item, Equals, "TEST")
 	c.Assert(exp.IsZero(), Not(Equals), true)
 
-	store.Delete("1")
+	cache.Delete("1")
 
-	c.Assert(store.Get("1"), Equals, nil)
+	c.Assert(cache.Get("1"), Equals, nil)
 
 	time.Sleep(time.Second / 8)
 
-	item, _ = store.GetWithExpiration("2")
+	item, _ = cache.GetWithExpiration("2")
 
-	c.Assert(store.Get("2"), Equals, nil)
+	c.Assert(cache.Get("2"), Equals, nil)
 	c.Assert(item, Equals, nil)
 
-	store.Flush()
+	cache.Flush()
 }
 
-func (s *CacheSuite) TestStoreWithoutJanitor(c *C) {
-	store := New(time.Second/32, 0)
+func (s *CacheSuite) TestCacheWithoutJanitor(c *C) {
+	cache := New(time.Second/32, 0)
 
-	store.Set("1", "TEST")
-	store.Set("2", "TEST")
-	store.Set("3", "TEST")
+	cache.Set("1", "TEST")
+	cache.Set("2", "TEST")
+	cache.Set("3", "TEST")
 
-	c.Assert(store.Has("2"), Equals, true)
-	c.Assert(store.Has("0"), Equals, false)
+	c.Assert(cache.Has("2"), Equals, true)
+	c.Assert(cache.Has("0"), Equals, false)
 
 	time.Sleep(time.Second / 16)
 
-	c.Assert(store.Has("1"), Equals, false)
-	c.Assert(store.Get("2"), Equals, nil)
+	c.Assert(cache.Expired(), Equals, 3)
+	c.Assert(cache.Has("1"), Equals, false)
+	c.Assert(cache.Get("2"), Equals, nil)
 
-	data, _ := store.GetWithExpiration("3")
+	data, _ := cache.GetWithExpiration("3")
 
 	c.Assert(data, Equals, nil)
 }
 
 func (s *CacheSuite) TestExpiration(c *C) {
-	store := New(time.Second/16, time.Minute)
+	cache := New(time.Second/16, time.Minute)
 
-	store.Set("1", "TEST")
+	cache.Set("1", "TEST")
 
 	time.Sleep(time.Second / 8)
 
-	item, _ := store.GetWithExpiration("1")
+	item, _ := cache.GetWithExpiration("1")
 	c.Assert(item, Equals, nil)
 
-	c.Assert(store.Get("1"), Equals, nil)
+	c.Assert(cache.Get("1"), Equals, nil)
 }
 
 func (s *CacheSuite) TestNil(c *C) {
-	var store *Store
+	var cache *Cache
 
-	store.Set("1", "TEST")
-	store.Delete("1")
-	store.Flush()
+	cache.Set("1", "TEST")
+	cache.Delete("1")
+	cache.Flush()
 
-	c.Assert(store.Get("1"), Equals, nil)
-	c.Assert(store.Has("1"), Equals, false)
+	c.Assert(cache.Size(), Equals, 0)
+	c.Assert(cache.Expired(), Equals, 0)
+	c.Assert(cache.Get("1"), Equals, nil)
+	c.Assert(cache.Has("1"), Equals, false)
 
-	item, exp := store.GetWithExpiration("1")
+	item, exp := cache.GetWithExpiration("1")
 	c.Assert(item, Equals, nil)
 	c.Assert(exp.IsZero(), Equals, true)
 }